Rename files to say work history instead of credentials
