+++ app/AndroidManifest.xml
@@ -567,7 +567,7 @@
             tools:node="merge" />
 
         <activity
-            android:name="org.commcare.activities.connect.PersonalIdCredentialActivity"/>
+            android:name="org.commcare.activities.connect.PersonalIdWorkHistoryActivity"/>
 
         <service
             android:name="org.commcare.services.CommCareFirebaseMessagingService"
