+++ app/res/drawable/ic_personalid_work_history_earned.xml
