+++ app/res/drawable/ic_personalid_work_history_ribbin.xml
