+++ app/res/layout/fragment_work_history_earned.xml
