+++ app/src/org/commcare/activities/connect/PersonalIdWorkHistoryActivity.kt
@@ -1,7 +1,6 @@
 package org.commcare.activities.connect
 
 import android.os.Bundle
-import android.util.Log
 import android.view.Menu
 import android.view.MenuItem
 import android.widget.Toast
@@ -9,7 +8,7 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.lifecycle.ViewModelProvider
 import com.google.android.material.tabs.TabLayoutMediator
 import org.commcare.activities.connect.viewmodel.PersonalIdCredentialViewModel
-import org.commcare.adapters.CredentialsViewPagerAdapter
+import org.commcare.adapters.WorkHistoryViewPagerAdapter
 import org.commcare.connect.network.connectId.PersonalIdApiErrorHandler
 import org.commcare.dalvik.R
 import org.commcare.dalvik.databinding.ActivityPersonalIdCredentialBinding
@@ -18,7 +17,7 @@ class PersonalIdCredentialActivity : AppCompatActivity() {
     private val binding: ActivityPersonalIdCredentialBinding by lazy {
         ActivityPersonalIdCredentialBinding.inflate(layoutInflater)
     }
-    private lateinit var credentialsViewPagerAdapter: CredentialsViewPagerAdapter
+    private lateinit var workHistoryViewPagerAdapter: WorkHistoryViewPagerAdapter
     private lateinit var personalIdCredentialViewModel: PersonalIdCredentialViewModel
     private var userName: String? = null
     private var profilePic: String? = null
@@ -33,7 +32,7 @@ class PersonalIdCredentialActivity : AppCompatActivity() {
         )[PersonalIdCredentialViewModel::class.java]
         userName = personalIdCredentialViewModel.userName
         profilePic = personalIdCredentialViewModel.profilePhoto
-        credentialsViewPagerAdapter = CredentialsViewPagerAdapter(this,userName!!,profilePic ?: "")
+        workHistoryViewPagerAdapter = WorkHistoryViewPagerAdapter(this,userName!!,profilePic ?: "")
         observeCredentialApiCall()
         fetchCredentialsFromNetwork()
         setUpUi()
@@ -45,7 +44,7 @@ class PersonalIdCredentialActivity : AppCompatActivity() {
             setDisplayShowHomeEnabled(true)
             setDisplayHomeAsUpEnabled(true)
         }
-        binding.vpCredentials.adapter = credentialsViewPagerAdapter
+        binding.vpCredentials.adapter = workHistoryViewPagerAdapter
         TabLayoutMediator(binding.tabCredentials, binding.vpCredentials) { tab, position ->
             tab.text = getString(titles[position])
             tab.setIcon(icons[position])
